+++ src/main/java/com/example/demo/entity/BookWithUser.java
@@ -1,5 +1,7 @@
 package com.example.demo.entity;
 
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
@@ -9,7 +11,8 @@ import java.util.Date;
 @TableName("bookwithuser")
 @Data
 public class BookWithUser {
-    private Integer id;
+
+    private Long id;
     private String isbn;
     private String bookName;
     private String nickName;
